refactor: Move get_strings_for_translation to api.py file
